# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: EduCollab API
  description: |
    API REST para a Plataforma de Aprendizado Colaborativo EduCollab.
    
    Esta API fornece endpoints para gerenciamento de usuários, cursos, anotações, 
    fórum e sistema de recomendações em uma plataforma educacional colaborativa.
    
    ## Funcionalidades Principais
    - Sistema de autenticação e autorização
    - Gerenciamento completo de usuários e perfis
    - CRUD de cursos e módulos educacionais
    - Sistema de anotações colaborativas
    - Fórum de discussões com comentários aninhados
    - Dashboard com estatísticas e métricas
    - Sistema de recomendações inteligentes
    
    ## Autenticação
    A API utiliza autenticação baseada em sessão. Para acessar endpoints protegidos,
    é necessário fazer login através do endpoint `/auth/login`.
    
  version: 1.0.0
  contact:
    name: EduCollab Support
    email: support@educollab.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Servidor de desenvolvimento local
  - url: https://api.educollab.com
    description: Servidor de produção
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Cadastro de novo usuário
      description: Cria uma nova conta de usuário na plataforma
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                  description: Nome completo do usuário
                  example: "João Silva"
                username:
                  type: string
                  description: Nome de usuário único (opcional, usa email se não fornecido)
                  example: "joao_silva"
                email:
                  type: string
                  format: email
                  description: Email único do usuário
                  example: "joao@educollab.com"
                password:
                  type: string
                  format: password
                  description: Senha do usuário (mínimo 6 caracteres)
                  example: "123456"
                role:
                  type: string
                  enum: [aluno, tutor, admin]
                  description: Papel do usuário no sistema
                  default: "aluno"
                  example: "aluno"
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuário criado com sucesso"
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos ou usuário já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login do usuário
      description: Autentica um usuário e inicia uma sessão
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: Email do usuário
                  example: "joao@educollab.com"
                password:
                  type: string
                  format: password
                  description: Senha do usuário
                  example: "123456"
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login realizado com sucesso"
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    get:
      tags:
        - Users
      summary: Listar usuários ativos
      description: Retorna uma lista de todos os usuários ativos do sistema
      responses:
        '200':
          description: Lista de usuários retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Obter usuário específico
      description: Retorna os dados de um usuário específico
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID único do usuário
          example: 1
      responses:
        '200':
          description: Dados do usuário retornados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Users
      summary: Atualizar usuário
      description: Atualiza os dados de um usuário específico
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID único do usuário
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "João Silva Santos"
                username:
                  type: string
                  example: "joao_santos"
                email:
                  type: string
                  format: email
                  example: "joao.santos@educollab.com"
                role:
                  type: string
                  enum: [aluno, tutor, admin]
                  example: "tutor"
                level:
                  type: string
                  example: "Cabo"
                avatar:
                  type: string
                  example: "https://example.com/avatar.jpg"
                password:
                  type: string
                  format: password
                  description: Nova senha (opcional)
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Users
      summary: Desativar usuário
      description: Desativa um usuário (soft delete)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID único do usuário
          example: 1
      responses:
        '200':
          description: Usuário desativado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuário desativado com sucesso"
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/stats:
    get:
      tags:
        - Statistics
      summary: Estatísticas do usuário
      description: Retorna estatísticas detalhadas de um usuário específico
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID único do usuário
          example: 1
      responses:
        '200':
          description: Estatísticas retornadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/recommendations:
    get:
      tags:
        - Statistics
      summary: Recomendações para o usuário
      description: Retorna cursos recomendados baseados no histórico do usuário
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID único do usuário
          example: 1
      responses:
        '200':
          description: Recomendações retornadas com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses:
    get:
      tags:
        - Courses
      summary: Listar cursos ativos
      description: Retorna uma lista de todos os cursos ativos
      responses:
        '200':
          description: Lista de cursos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Courses
      summary: Criar novo curso
      description: Cria um novo curso na plataforma
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - category
                - instructor_id
              properties:
                title:
                  type: string
                  description: Título do curso
                  example: "React Fundamentals"
                description:
                  type: string
                  description: Descrição detalhada do curso
                  example: "Aprenda os conceitos fundamentais do React"
                category:
                  type: string
                  description: Categoria do curso
                  example: "Desenvolvimento Web"
                difficulty:
                  type: string
                  enum: [iniciante, intermediario, avancado]
                  description: Nível de dificuldade
                  default: "iniciante"
                  example: "iniciante"
                duration_hours:
                  type: integer
                  description: Duração em horas
                  default: 0
                  example: 20
                instructor_id:
                  type: integer
                  description: ID do instrutor responsável
                  example: 2
      responses:
        '201':
          description: Curso criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{courseId}:
    get:
      tags:
        - Courses
      summary: Obter curso específico
      description: Retorna os dados de um curso específico
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
          description: ID único do curso
          example: 1
      responses:
        '200':
          description: Dados do curso retornados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          description: Curso não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Courses
      summary: Atualizar curso
      description: Atualiza os dados de um curso específico
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
          description: ID único do curso
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "React Avançado"
                description:
                  type: string
                  example: "Conceitos avançados do React"
                category:
                  type: string
                  example: "Desenvolvimento Web"
                difficulty:
                  type: string
                  enum: [iniciante, intermediario, avancado]
                  example: "avancado"
                duration_hours:
                  type: integer
                  example: 40
      responses:
        '200':
          description: Curso atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '404':
          description: Curso não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Courses
      summary: Desativar curso
      description: Desativa um curso (soft delete)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
          description: ID único do curso
          example: 1
      responses:
        '200':
          description: Curso desativado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Curso desativado com sucesso"
        '404':
          description: Curso não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /enrollments:
    post:
      tags:
        - Enrollments
      summary: Criar inscrição
      description: Inscreve um usuário em um curso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - course_id
              properties:
                user_id:
                  type: integer
                  description: ID do usuário
                  example: 1
                course_id:
                  type: integer
                  description: ID do curso
                  example: 1
      responses:
        '201':
          description: Inscrição criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '400':
          description: Usuário já inscrito ou dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/enrollments:
    get:
      tags:
        - Enrollments
      summary: Listar inscrições do usuário
      description: Retorna todas as inscrições ativas de um usuário
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID único do usuário
          example: 1
      responses:
        '200':
          description: Lista de inscrições retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Enrollment'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /notes/{noteId}:
    put:
      tags:
        - Notes
      summary: Atualizar anotação
      description: Atualiza uma anotação específica
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
          description: ID único da anotação
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "React Hooks Avançados"
                content:
                  type: string
                  example: "Conceitos avançados de React Hooks..."
                tags:
                  type: string
                  example: '["react", "hooks", "avancado"]'
                is_shared:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Anotação atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Anotação não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: ID único do usuário
          example: 1
        name:
          type: string
          description: Nome completo do usuário
          example: "João Silva"
        username:
          type: string
          description: Nome de usuário único
          example: "joao_silva"
        email:
          type: string
          format: email
          description: Email único do usuário
          example: "joao@educollab.com"
        role:
          type: string
          enum: [aluno, tutor, admin]
          description: Papel do usuário no sistema
          example: "aluno"
        level:
          type: string
          description: Nível militar do usuário
          example: "Recruta"
        avatar:
          type: string
          nullable: true
          description: URL do avatar do usuário
          example: null
        created_at:
          type: string
          format: date-time
          description: Data de criação da conta
          example: "2025-06-30T13:10:05.482434"
        last_login:
          type: string
          format: date-time
          nullable: true
          description: Data do último login
          example: null
        is_active:
          type: boolean
          description: Se o usuário está ativo
          example: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: "Credenciais inválidas"